+++ src/ja/mangasaison/res/mipmap-mdpi/ic_launcher.png
